+++ my-quizz-builder-api/src/main/java/co/simplon/myquizzbuilder/services/AvailableAnswerService.java
@@ -4,5 +4,5 @@ import co.simplon.myquizzbuilder.dtos.quiz.AvailableAnswerCreateDto;
 
 public interface AvailableAnswerService {
     void create(AvailableAnswerCreateDto inputs,
-	    Integer questionId);
+	    Long questionId);
 }
